File organization
also added inlining attributes to interpolators
