+++ Runtime/Easing Functions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 87bb609e127ac4530bd1b41b2bbf3176
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
