+++ Runtime/Easing Functions/EasingFunction+Back.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2a42f9b34815e465299f2c89ec3e5969
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
