+++ Runtime/Easing Functions/EasingFunction+Bounce.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2ed45aec82d6b465d92714d79cb86a44
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
