+++ Runtime/Easing Functions/EasingFunction+Circular.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 28d316b732170462d8b1ebc61fa2b2f4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
