+++ Runtime/Easing Functions/EasingFunction+Cubic.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 95ed7edbc264a4b03b87882dbaa33a27
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
