+++ Runtime/Easing Functions/EasingFunction+Elastic.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1149000d162b64c4e9ed77af5b31e1bd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
