+++ Runtime/Easing Functions/EasingFunction+Exponential.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ab3e258bf598046b6a47138132f7dd6d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
