+++ Runtime/Easing Functions/EasingFunction+Linear.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b82b284eb64c84529b9ed5d93afd59d2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
