+++ Runtime/Easing Functions/EasingFunction+Quadratic.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1874d15a95c684789b6d48ae927c4234
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
