+++ Runtime/Easing Functions/EasingFunction+Quartic.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2606308c5f7bd4412ac552cb9b62a38c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
