+++ Runtime/Easing Functions/EasingFunction+Quintic.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a0620d606ea1d4ac7a5cc65a6c4c23a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
