+++ Runtime/Easing Functions/EasingFunction+Sinusoidal.cs
@@ -1,0 +1,29 @@
+// Developed With Love by Ryan Boyer http://ryanjboyer.com <3
+
+using System.Runtime.CompilerServices;
+
+namespace EaseKit {
+	using static EasingUtility;
+
+	public static partial class EasingFunction {
+		public static class Sinusoidal {
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			public static float EaseOut(float t, float b, float c, float d) => c * sin(t / d * (PI * 0.5f)) + b;
+
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			public static float EaseIn(float t, float b, float c, float d) => -c * cos(t / d * (PI * 0.5f)) + c + b;
+
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			public static float EaseInOut(float t, float b, float c, float d)
+				=> c * 0.5f * (1 - cos(t / d * PI)) + b;
+
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			public static float EaseOutIn(float t, float b, float c, float d) {
+				if (t < d * 0.5f) {
+					return EaseOut(t * 2, b, c * 0.5f, d);
+				}
+				return EaseIn((t * 2) - d, b + c * 0.5f, c * 0.5f, d);
+			}
+		}
+	}
+}
