+++ Runtime/Easing Functions/EasingFunction+Sinusoidal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 013063d1bcae544c7942be92bc2140cc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
