+++ Runtime/Easing Utility/EasingUtility+Easing.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d215662b039454246bb8f3ed5e84fce2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
