+++ Runtime/Easing Utility/EasingUtility+Evaluate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9257b182c474c4a44a027c668d7220b2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
