+++ Runtime/Easing Utility/EasingUtility+math.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0b084140890b7447db006ce8253c4907
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
