+++ Runtime/Interpolators/Unity Mathematics/quaternionInterpolator.cs
@@ -1,12 +1,16 @@
+// Developed With Love by Ryan Boyer http://ryanjboyer.com <3
+
 #if UNITY_MATHEMATICS
+using System.Runtime.CompilerServices;
 using Unity.Mathematics;
 
 namespace EaseKit {
-    public readonly struct quaternionInterpolator : IInterpolator<quaternion> {
-        public static readonly quaternionInterpolator shared = new quaternionInterpolator();
+	public readonly struct quaternionInterpolator : IInterpolator<quaternion> {
+		public static readonly quaternionInterpolator shared = new quaternionInterpolator();
 
-        public quaternion Evaluate(in quaternion start, in quaternion end, float percent)
-            => math.slerp(start, end, percent);
-    }
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public quaternion Evaluate(in quaternion start, in quaternion end, float percent)
+			=> math.slerp(start, end, percent);
+	}
 }
 #endif
