+++ Runtime/Interpolators/UnityEngine/ColorInterpolator.cs
@@ -1,10 +1,14 @@
+// Developed With Love by Ryan Boyer http://ryanjboyer.com <3
+
+using System.Runtime.CompilerServices;
 using UnityEngine;
 
 namespace EaseKit {
-    public readonly struct ColorInterpolator : IInterpolator<Color> {
-        public static readonly ColorInterpolator shared = new ColorInterpolator();
+	public readonly struct ColorInterpolator : IInterpolator<Color> {
+		public static readonly ColorInterpolator shared = new ColorInterpolator();
 
-        public Color Evaluate(in Color start, in Color end, float percent)
-            => Color.Lerp(start, end, percent);
-    }
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public Color Evaluate(in Color start, in Color end, float percent)
+			=> Color.Lerp(start, end, percent);
+	}
 }
