+++ Runtime/Interpolators/UnityEngine/Vector3IntInterpolator.cs
@@ -1,9 +1,13 @@
+// Developed With Love by Ryan Boyer http://ryanjboyer.com <3
+
+using System.Runtime.CompilerServices;
 using UnityEngine;
 
 namespace EaseKit {
 	public readonly struct Vector3IntInterpolator : IInterpolator<Vector3Int> {
 		public static readonly Vector3IntInterpolator shared = new Vector3IntInterpolator();
 
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public Vector3Int Evaluate(in Vector3Int start, in Vector3Int end, float percent)
 			=> new Vector3Int(
 				intInterpolator.shared.Evaluate(start.x, end.x, percent),
