+++ Runtime/Interpolators/floatInterpolator.cs
@@ -1,8 +1,13 @@
-namespace EaseKit {
-    public readonly struct floatInterpolator : IInterpolator<float> {
-        public static readonly floatInterpolator shared = new floatInterpolator();
+// Developed With Love by Ryan Boyer http://ryanjboyer.com <3
 
-        public float Evaluate(in float start, in float end, float percent)
-            => EasingUtility.lerp(start, end, percent);
-    }
+using System.Runtime.CompilerServices;
+
+namespace EaseKit {
+	public readonly struct floatInterpolator : IInterpolator<float> {
+		public static readonly floatInterpolator shared = new floatInterpolator();
+
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public float Evaluate(in float start, in float end, float percent)
+			=> EasingUtility.lerp(start, end, percent);
+	}
 }
